% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_reg.R
\name{knots_quantile}
\alias{knots_quantile}
\title{Generating interior knots sequence for given data points}
\usage{
knots_quantile(x, dimension, degree = 3)
}
\arguments{
\item{x}{Numeric vector representing data points}

\item{dimension}{Numeric value indicating the number of basis functions}

\item{degree}{Numeric value representing the spline degree}
}
\value{
Numeric vector with interior knots sequence
}
\description{
This function computes a knot sequence based on the quantiles of the data points
}
\examples{
x = runif(50, 0, 1)
knots = knots_quantile(x, 5, 3)
print(knots)
}
